+++ packages/dci/RequestsListItem.js
@@ -6,7 +6,7 @@ import { property, state, customElement } from 'lit/decorators.js';
 import { choose } from 'lit/directives/choose.js';
 import { when } from 'lit/directives/when.js';
 import { DCIStyles } from './styles.js';
-import { getPendingDemand, approveDemand } from './utils/data-consumer-api.js';
+import { getPendingDemand, approveDemand, denyDemand } from './utils/data-consumer-api.js';
 
 var REQ_ITEM_UI_STATE;
 (function (REQ_ITEM_UI_STATE) {
@@ -31,6 +31,7 @@ let RequestsListItem = class RequestsListItem extends LitElement {
         };
         this._open = false;
         this._uiState = REQ_ITEM_UI_STATE.PENDING_DECISION;
+        this._message = '';
     }
     willUpdate(_changedProperties) {
         if (_changedProperties.has('demand')) {
@@ -39,13 +40,19 @@ let RequestsListItem = class RequestsListItem extends LitElement {
             });
         }
     }
-    approveDemand() {
-        approveDemand(this.demand.id, 'Approved').then(() => {
+    handleApproveDemandClick() {
+        approveDemand(this.demand.id, this._message).then(() => {
             this._uiState = REQ_ITEM_UI_STATE.APPROVED;
         });
     }
-    denyDemand() {
-        this._uiState = REQ_ITEM_UI_STATE.DENIED;
+    handleDenyDemandClick() {
+        denyDemand(this.demand.id, this._message).then(() => {
+            this._uiState = REQ_ITEM_UI_STATE.DENIED;
+        });
+    }
+    handleMessageInput(e) {
+        const { value } = e.target;
+        this._message = value;
     }
     render() {
         return html `
@@ -66,22 +73,35 @@ let RequestsListItem = class RequestsListItem extends LitElement {
               <span class="dmd-info-element">${this.demand.action}</span>
             </div>
             ${when(this._open, () => html `
-                <div id="dmd-details-ctr">
+                <div id="dmd-response-ctr">
                   ${when(this._demandDetails, () => html `
-                      <button
-                        id="approve-btn"
-                        class="dmd-btn animated-btn"
-                        @click=${this.approveDemand}
-                      >
-                        Approve
-                      </button>
-                      <button
-                        id="deny-btn"
-                        class="dmd-btn animated-btn"
-                        @click=${this.denyDemand}
-                      >
-                        Deny
-                      </button>
+                      <div id="response-msg-ctr">
+                        <label for="response-msg"
+                          >${msg('Optional Message')}</label
+                        >
+                        <textarea
+                          id="response-msg"
+                          rows="5"
+                          cols="50"
+                          @input=${this.handleMessageInput}
+                        ></textarea>
+                      </div>
+                      <div id="response-btns-ctr">
+                        <button
+                          id="approve-btn"
+                          class="dmd-btn animated-btn"
+                          @click=${this.handleApproveDemandClick}
+                        >
+                          Approve
+                        </button>
+                        <button
+                          id="deny-btn"
+                          class="dmd-btn animated-btn"
+                          @click=${this.handleDenyDemandClick}
+                        >
+                          Deny
+                        </button>
+                      </div>
                     `, () => html ` Getting demand details... `)}
                 </div>
               `)}
@@ -126,11 +146,34 @@ RequestsListItem.styles = [
         align-items: center;
       }
 
-      #dmd-details-ctr {
+      #dmd-response-ctr {
+        display: grid;
+        justify-content: center;
+        padding: 20px;
+        row-gap: 20px;
+      }
+
+      #response-msg-ctr {
+        display: grid;
+        row-gap: 10px;
+      }
+
+      #response-msg-ctr label {
+        display: block;
+        text-align: left;
+      }
+
+      #response-msg {
+        display: block;
+        background: #f8f8fc;
+        border: 1px solid #d9d9d9;
+        border-radius: 8px;
+      }
+
+      #response-btns-ctr {
         display: flex;
         justify-items: center;
         justify-content: center;
-        padding: 20px;
         column-gap: 40px;
       }
 
@@ -151,7 +194,7 @@ RequestsListItem.styles = [
         border-radius: 10px;
         border-width: 4px;
         background: none;
-        width: 200px;
+        width: 150px;
         text-align: center;
       }
 
@@ -177,6 +220,9 @@ __decorate([
 __decorate([
     state()
 ], RequestsListItem.prototype, "_uiState", void 0);
+__decorate([
+    state()
+], RequestsListItem.prototype, "_message", void 0);
 RequestsListItem = __decorate([
     customElement('requests-list-item')
 ], RequestsListItem);