+++ demos/static/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/blindnet-io/privacy-components-web/compare/v0.5.0...v0.5.1) (2022-09-02)
+
+**Note:** Version bump only for package @blindnet-demos/static
+
+
+
+
+
+# [0.5.0](https://github.com/blindnet-io/privacy-components-web/compare/v0.4.0...v0.5.0) (2022-09-02)
+
+**Note:** Version bump only for package @blindnet-demos/static
+
+
+
+
+
 # [0.4.0](https://github.com/blindnet-io/privacy-components-web/compare/v0.4.0-beta.7...v0.4.0) (2022-08-21)
 
 **Note:** Version bump only for package @blindnet-demos/static